Rename IDs in a single test case
This is a bit stupid, but essentially lxml / libxml2 will throw a
shit-fit when it tries to load a document with two elements that have
the same `xml:id`.  I wasn't able to convince it to _not_ do this, so
rather than try and fix it somehow, we're just editing the test case.
XML is (occasionally) the worst.